Updates
Added "Rift.Runtime.Logging" module to the core.
Basic input management (will be entirely moved to a separate module asap)
Added hardcoded platform checks for the StorageSystem.
Created FNV-1a implementation for hashing strings - used in input management callbacks.
Made sure that explicit casting to float is done in the AbsoluteVec functions, part of the Vector classes.
Added new "PreEngineInit" function to the Platform abstractization and a new Thread factory function.
Modified return type of Bitmap class to be a constant reference of a color vector in order to prevent the creation of a copy of the pixel data.
Modified docs for the Color class to accurately specify that the data is packed in RGBA format.
Added cyan color to the list.
Modified IShaderProgram's AddShader parameter to represent a unique smart pointer of a IShader rather than a direct IShader pointer.
Created a IThread interface class for thread abstractization.
Removed a comment and added a new function to fetch the device context of a window in IWindow.